(updates) Working accounts/transactions

+++ go/backend/models/account.go
@@ -1,6 +1,6 @@
 package models
 
 type Account struct {
-	ID   uint   `json:"id" gorm:"primary_key"`
+	ID   int64  `json:"id" gorm:"primary_key"`
 	Name string `json:"name"`
 }
